backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
sqlalchemy
python-jose
passlib[bcrypt]
python-multipart